Create README.md
This machine learning project "Dami's url predictor" is a uniform resource locator (url) prediction system. It classifies with a high level of accuracy if urls are likely malicious (including spam/phishing etc) websites or if they are harmless/"benign".  It also gives users the option to view urls that might be similar/belong to the same person or company based on users desired similarity threshold. The threshold values range from 0.1-1. This model also gives the users the option to have this saved automatically to an excel file or printed.